add state schema for state + transactional state
