+++ stackstate_checks_base/stackstate_checks/base/checks/__init__.py
@@ -8,7 +8,7 @@ try:
     from checks.network_checks import NetworkCheck, Status, EventType
 except ImportError:
     from .base import AgentCheck, TopologyInstance, StackPackInstance, AgentIntegrationInstance
-    from .base_v2 import AgentCheckV2, Stateful, Transactional, StatefulAgentCheck, TransactionalAgentCheck
+    from .v2 import AgentCheckV2, StatefulMixin, TransactionalMixin, StatefulAgentCheck, TransactionalAgentCheck
     from ..utils.health_api import HealthStream, HealthStreamUrn, Health
     from .network import NetworkCheck, Status, EventType
 
@@ -24,8 +24,8 @@ __all__ = [
     'HealthStreamUrn',
     'Health',
     'AgentCheckV2',
-    'Stateful',
-    'Transactional',
+    'StatefulMixin',
+    'TransactionalMixin',
     'StatefulAgentCheck',
     'TransactionalAgentCheck'
 ]
