+++ stackstate_checks_base/stackstate_checks/base/checks/v2/check_error.py
@@ -1,0 +1,10 @@
+
+
+class CheckError(Exception):
+    """
+    Agent error that needs to be converted to Critical Service Check
+    """
+
+    def to_string(self):
+        # type: () -> str
+        return "check error: {}".format(self.message)
