+++ stackstate_checks_base/stackstate_checks/base/checks/v2/instance_type.py
@@ -1,0 +1,4 @@
+from schematics import Model
+from typing import Any, Dict, TypeVar
+
+InstanceType = TypeVar('InstanceType', Model, Dict[str, Any])
